requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
requests>=2.28
